Pull out the ignore method which shouldn't really be there.
